refactor(document-processor): move extractMaxPageNumber to utils and add dedicated tests
Extract `extractMaxPageNumber` functionality from `document-processor` to a new utility module `toc-markdown-utils` for better modularity and reusability. Add comprehensive tests in `toc-markdown-utils.test.ts` to validate behavior. Update references and remove the old implementation in `document-processor`.
