+++ packages/document-processor/src/utils/index.ts
@@ -2,3 +2,4 @@ export { RefResolver } from './ref-resolver';
 export { IdGenerator } from './id-generator';
 export { TextCleaner } from './text-cleaner';
 export { MarkdownConverter } from './markdown-converter';
+export { extractMaxPageNumber } from './toc-markdown-utils';
